% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{attachPackages}
\alias{attachPackages}
\title{Require, and first install from CRAN if necessary, a list of packages.}
\usage{
attachPackages(...)
}
\arguments{
\item{...}{Package names as strings.}
}
\description{
This function will load packages to your project using the require function.
First, however, it will try to install them from CRAN. Thus, you can use this
function to install and require packages in one go, so as not to worry about
which packages you do or do not have installed on your system. Note that
installing from CRAN means that packages will not be installed if they do not
exist on CRAN. However, if they're already installed on your system, they
will still be loaded.
}
\examples{
attachPackages("data.table", "pipeR")

}

