% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{starMatrix}
\alias{starMatrix}
\title{Generated a correlation matrix with significance stars (*)}
\usage{
starMatrix(vars, ...)
}
\arguments{
\item{vars}{A matrix or dataframe}

\item{...}{Additional parameters to pass to psych::corr.test()}
}
\value{
Correlation matrix with significance stars
}
\description{
This function correlates the columns of a matrix or dataframe, and returns
the results with stars indicating their level of statistical significance.
The stars are used to indicate the following significance levels: *** p <
.001, ** p < .01, * p < .05
}
\examples{
starMatrix(iris[, -5])
}

